test(rust): move large unit tests out of implementation files
